% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarDFToPlot.R
\name{BarDFToPlot}
\alias{BarDFToPlot}
\title{Helper function to convert a Dataframe into a Bar Plot}
\usage{
BarDFToPlot(Ds_df, Individual_Name, hglData = NULL)
}
\arguments{
\item{The}{data frame, individual names, and highlighted data (defaults to NULL)}
}
\value{
A bar plot
}
\description{
Helper function to convert a Dataframe into a Bar Plot
}
\examples{
BarDFToPlot(Dataframe, Names, NULL)
}
