% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barLegendInd.R
\name{barLegendInd}
\alias{barLegendInd}
\title{Helper function to generate a Legend, provided a K-value}
\usage{
barLegendInd(kVal)
}
\arguments{
\item{The}{K-value}
}
\value{
A legend object
}
\description{
Helper function to generate a Legend, provided a K-value
}
\examples{
barLegendInd(K-Value)
}
