% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDBS_To_Heat.R
\name{BDBS_To_Heat}
\alias{BDBS_To_Heat}
\title{Helper function to generate a heatmap}
\usage{
BDBS_To_Heat(loadedMat, popName, procs, iNames, slctInt = NULL,
  heatZmLvl = NULL)
}
\arguments{
\item{The}{loaded matrix, population file name, process that has been selected, the individual names, the selected value, the zoom level (now obsolete)}
}
\value{
An object that is used to generate the heatmap
}
\description{
Helper function to generate a heatmap
}
\examples{
BDBS_To_Heat(matrix, population name, process, individual names, selected value)
}
