% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPopDF.R
\name{boxPopDF}
\alias{boxPopDF}
\title{Function to convert a vector list into a data frame}
\usage{
boxPopDF(loadMatList)
}
\arguments{
\item{The}{List}
}
\value{
A dataframe
}
\description{
Function to convert a vector list into a data frame
}
\examples{
boxPopDF(list)
}
