% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleData.R
\name{circleData}
\alias{circleData}
\title{Helper function for MDS Plots}
\usage{
circleData(xMDS, yMDS)
}
\arguments{
\item{The}{x and y for the MDS Plot}
}
\value{
The points on the plot
}
\description{
Helper function for MDS Plots
}
\examples{
circleData(x, y)
}
