% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlePts.R
\name{circlePts}
\alias{circlePts}
\title{Helper function for MDS Plots}
\usage{
circlePts(xCent, yCent, radius, pts = 100)
}
\arguments{
\item{x, }{y, the radius, and the number of points.}
}
\value{
Data on the MDS Plot
}
\description{
Helper function for MDS Plots
}
\examples{
circlePts(x,y, radius = r, pts = z)
}
