% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complKinMat.R
\name{complKinMat}
\alias{complKinMat}
\title{Function that creates the kinship matrix}
\usage{
complKinMat(kinMatData, popName)
}
\arguments{
\item{Kinship}{Matrix Data, and the Population Name}
}
\value{
A new matrix
}
\description{
Function that creates the kinship matrix
}
\examples{
complKinMat(matrix, CLM)
}
