% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covToDist.R
\name{covToDist}
\alias{covToDist}
\title{Function to convert covariance to distance}
\usage{
covToDist(covMat)
}
\arguments{
\item{Covariance}{matrix}
}
\value{
Covariance Distance Matrix
}
\description{
Function to convert covariance to distance
}
\examples{
covToDis(CovarianceMatrix)
}
