% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covToPlot.R
\name{covToPlot}
\alias{covToPlot}
\title{Function to convert covariance matrix into a plot}
\usage{
covToPlot(covMatrix, hglData = NULL)
}
\arguments{
\item{The}{covariance matrix, and any highlighted data}
}
\value{
The plot
}
\description{
Function to convert covariance matrix into a plot
}
\examples{
covToPlot(Matrix, Highlighted Data)
}
