% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFToBarPlot.R
\name{DFToBarPlot}
\alias{DFToBarPlot}
\title{Helper function to convert a data frame into a bar plot}
\usage{
DFToBarPlot(toPlot_DF, highlightData = NULL, Population_Name,
  isName = FALSE)
}
\arguments{
\item{The}{data frame to be plotted, highlighted data, population name, and a name parameter (defaults to FALSE)}
}
\value{
A bar plot
}
\description{
Helper function to convert a data frame into a bar plot
}
\examples{
DFToBarPlot(Dataframe, HighlightedData, Population)
}
