% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFToBoxPlot.R
\name{DFToBoxPlot}
\alias{DFToBoxPlot}
\title{Helper function to convert a data frame into a box plot}
\usage{
DFToBoxPlot(IBS_df, ordByMed)
}
\arguments{
\item{The}{data frame to be plotted, and the means by which the plot will be ordered}
}
\value{
A box plot
}
\description{
Helper function to convert a data frame into a box plot
}
\examples{
DFToBoxPlot(Dataframe, Ordering)
}
