% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFToMDSPlot.R
\name{DFToMDSPlot}
\alias{DFToMDSPlot}
\title{Helper function to convert a data frame into a MDS plot}
\usage{
DFToMDSPlot(MDS_DF, circlePoints, Label_Name, eigVals, slctVals = NULL)
}
\arguments{
\item{The}{data frame to be plotted, the points on the plot, label names, eigenvalues, and the selected values}
}
\value{
An MDS plot
}
\description{
Helper function to convert a data frame into a MDS plot
}
\examples{
DFToMDSPlot(MDS Dataframe, Points list, Name of Label, Eigenvalues, Selected Values)
}
