% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ds_dfToPlot.R
\name{Ds_dfToPlot}
\alias{Ds_dfToPlot}
\title{Helper function to convert a data frame into a plot}
\usage{
Ds_dfToPlot(Ds, Vs, plotType, maxKInd)
}
\arguments{
\item{Ds, }{Vs, type of plot, k-value}
}
\value{
Plot
}
\description{
Helper function to convert a data frame into a plot
}
\examples{
Ds_dfToPlot(Ds, Vs, plotType, K value)
}
