% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsToDataFrame.R
\name{DsToDataFrame}
\alias{DsToDataFrame}
\title{Helper function to convert a DS to a Data Frame}
\usage{
DsToDataFrame(selectedDs, maxK = 9)
}
\arguments{
\item{Selected}{DS, and the max K value (defaults to 9)}
}
\value{
A data frame
}
\description{
Helper function to convert a DS to a Data Frame
}
\examples{
DSToDataFrame(Selected Ds, maximum K)
}
