% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsToVs.R
\name{DsToVs}
\alias{DsToVs}
\title{Helper function to convert a DS into a normalized Vector}
\usage{
DsToVs(Ds, maxVal)
}
\arguments{
\item{The}{DS, and the maximum Value}
}
\value{
A normalized vector
}
\description{
Helper function to convert a DS into a normalized Vector
}
\examples{
DSToVS(DS, Maximum value)
}
