% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillNA.R
\name{fillNA}
\alias{fillNA}
\title{Helper function to replace NA values with 'C'}
\usage{
fillNA(matInp, maxValC)
}
\arguments{
\item{The}{input matrix, and the maximum value}
}
\value{
The modified input matrix
}
\description{
Helper function to replace NA values with 'C'
}
\examples{
fillNA(input matrix, maximum value)
}
