% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnameToIndDF.R
\name{fnameToIndDF}
\alias{fnameToIndDF}
\title{Helper function to convert a file into a data frame of individuals}
\usage{
fnameToIndDF(loadedMat, popFName, indNames, plotType, origInd = NULL,
  maxKValInd = NULL, prcs = NULL, slctNum = NULL, heatZmLvl = NULL)
}
\arguments{
\item{The}{loaded matrix, population file name, individual names, type of plot, origInd, maximum value individuals, the selected process, the number of selected, and the zoom of the heatmap}
}
\value{
A data frame of the individuals
}
\description{
Helper function to convert a file into a data frame of individuals
}
\examples{
fnameToIndDF(matrix, filename, individuals names, type of plot)
}
