% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnameToPopDF.R
\name{fnameToPopDF}
\alias{fnameToPopDF}
\title{Helper function to convert a file into a data frame of populations}
\usage{
fnameToPopDF(loadedList, maxKVal = NULL, plotType = NULL)
}
\arguments{
\item{The}{loaded matrix, max K value, and the type of plot}
}
\value{
A data frame of the individuals
}
\description{
Helper function to convert a file into a data frame of populations
}
\examples{
fnameToPopDF(matrix, k-value, type of plot)
}
