% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileNames.R
\name{getFileNames}
\alias{getFileNames}
\title{Helper function to get filenames from data directory}
\usage{
getFileNames()
}
\arguments{
\item{No}{inputparamters}
}
\value{
List of file names
}
\description{
Helper function to get filenames from data directory
}
\examples{
getFileNames()
}
