% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndNames.R
\name{getIndNames}
\alias{getIndNames}
\title{Helper function to get the names of individuals}
\usage{
getIndNames(fileName)
}
\arguments{
\item{The}{filename of the population}
}
\value{
A list of individual names
}
\description{
Helper function to get the names of individuals
}
\examples{
getIndNames(filename)
}
