% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopulationName.R
\name{getPopulationName}
\alias{getPopulationName}
\title{Helper function to get population name from a file}
\usage{
getPopulationName(fileName)
}
\arguments{
\item{The}{filename containing population data}
}
\value{
The population name
}
\description{
Helper function to get population name from a file
}
\examples{
getPopulationName(filename)
}
