% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessName.R
\name{getProcessName}
\alias{getProcessName}
\title{Helper function to get the process name from an input file}
\usage{
getProcessName(fileName)
}
\arguments{
\item{File}{name}
}
\value{
Process name
}
\description{
Helper function to get the process name from an input file
}
\examples{
getProcessName(Filename)
}
