% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hBN_To_Dend.R
\name{hBN_To_Dend}
\alias{hBN_To_Dend}
\title{Helper function to convert hBN to a Dendrogram}
\usage{
hBN_To_Dend(loadedMat, METHOD = NULL, K = NULL)
}
\arguments{
\item{Loaded}{Matrix, the method, and the k-value}
}
\value{
Plot and object to select from
}
\description{
Helper function to convert hBN to a Dendrogram
}
\examples{
hBN_To_Dend(matrix, method, k-value)
}
