% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hBNToMDSDF.R
\name{hBNToMDSDF}
\alias{hBNToMDSDF}
\title{Helper function to convert hBN to the MDS Dataframe}
\usage{
hBNToMDSDF(loadedList)
}
\arguments{
\item{Loaded}{list}
}
\value{
A data frame
}
\description{
Helper function to convert hBN to the MDS Dataframe
}
\examples{
hBNToMDSDF(List)
}
