% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatSlctSize.R
\name{heatSlctSize}
\alias{heatSlctSize}
\title{OBSOLETE - Helper function to help zoom in heatmap}
\usage{
heatSlctSize(heatZoom, numInds)
}
\arguments{
\item{Current}{zoom, number of individuals}
}
\value{
Selected size
}
\description{
OBSOLETE - Helper function to help zoom in heatmap
}
\examples{
heatSlctSize(Zoom, Number of Individuals)
}
