% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hIBSMsTo0To1.R
\name{hIBSMsTo0To1}
\alias{hIBSMsTo0To1}
\title{Helper function to get min and max of the matrix}
\usage{
hIBSMsTo0To1(hIBSMs)
}
\arguments{
\item{hIBS}{Matrix}
}
\value{
Matrix
}
\description{
Helper function to get min and max of the matrix
}
\examples{
hIBSMsTo0To1(Matrix)
}
