% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMaxVal.R
\name{listMaxVal}
\alias{listMaxVal}
\title{Helper function to get maximum value}
\usage{
listMaxVal(listInput)
}
\arguments{
\item{List}{of values}
}
\value{
Maximum Value in list
}
\description{
Helper function to get maximum value
}
\examples{
listMaxVal(List)
}
