% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVecRel.R
\name{listVecRel}
\alias{listVecRel}
\title{Helper function to get data names}
\usage{
listVecRel(inputVec, dispVec, dispVecList)
}
\arguments{
\item{Input}{vector, the displaced vector, the vector list}
}
\value{
Names
}
\description{
Helper function to get data names
}
\examples{
listVecRel(input vector, displayed vector, list)
}
