% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMultiPopData.R
\name{loadMultiPopData}
\alias{loadMultiPopData}
\title{Helper function to load multiple population data}
\usage{
loadMultiPopData(fileNameVec, fileDPVec, prcsName, extn)
}
\arguments{
\item{File}{names vector, data paths vector, process name, file extension}
}
\value{
Matrix with data
}
\description{
Helper function to load multiple population data
}
\examples{
loadMultiPopData(file names, data paths, process, extension)
}
