% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matToDFMDS.R
\name{matToDFMDS}
\alias{matToDFMDS}
\title{Helper function to convert matrix to MDS data frame}
\usage{
matToDFMDS(Ds, iNames)
}
\arguments{
\item{Data, }{individual names}
}
\value{
List with values for MDS
}
\description{
Helper function to convert matrix to MDS data frame
}
\examples{
matToDFMDS(DS, Names)
}
