% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matToUpTriVec.R
\name{matToUpTriVec}
\alias{matToUpTriVec}
\title{Helper function to get true values of upper triangular matrix}
\usage{
matToUpTriVec(matElm)
}
\arguments{
\item{elements}{of the matrix}
}
\value{
upper triangular matrix values
}
\description{
Helper function to get true values of upper triangular matrix
}
\examples{
matToUpTriVec(elements)
}
