% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxNZeroIndx.R
\name{maxNZeroIndx}
\alias{maxNZeroIndx}
\title{Helper function to find maximum non-zero index in a vector}
\usage{
maxNZeroIndx(vec)
}
\arguments{
\item{Vector}{in which to find max index}
}
\value{
Maximum index
}
\description{
Helper function to find maximum non-zero index in a vector
}
\examples{
maxNZeroIndx(vector)
}
