% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsToDs.R
\name{MsToDs}
\alias{MsToDs}
\title{Helper function to convert MS matrix to DS matrix}
\usage{
MsToDs(Ms, toChange = FALSE)
}
\arguments{
\item{MS, }{toChange boolean}
}
\value{
DS
}
\description{
Helper function to convert MS matrix to DS matrix
}
\examples{
MSToDS(MS)
}
