% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCovToDist.R
\name{multiCovToDist}
\alias{multiCovToDist}
\title{Helper function to convert covariance matrices to distance}
\usage{
multiCovToDist(multiCovMat)
}
\arguments{
\item{Covariance}{matrices}
}
\value{
List of covariance
}
\description{
Helper function to convert covariance matrices to distance
}
\examples{
multiCovToDist(covariance matrices)
}
