% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDsToVs.R
\name{multiDsToVs}
\alias{multiDsToVs}
\title{Helper function to convert matrix list to vector list}
\usage{
multiDsToVs(multiDsMat)
}
\arguments{
\item{Matrix}{list}
}
\value{
Vector list
}
\description{
Helper function to convert matrix list to vector list
}
\examples{
multiDsToVs(matrix list)
}
