% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihIBSMsTo0To1.R
\name{multihIBSMsTo0To1}
\alias{multihIBSMsTo0To1}
\title{Helper function}
\usage{
multihIBSMsTo0To1(hIBSMsList)
}
\arguments{
\item{hIBS}{Matrix List}
}
\value{
Modified List
}
\description{
Helper function
}
\examples{
multihIBSMsTo0To1(List of matrices)
}
