% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiMsToDs.R
\name{multiMsToDs}
\alias{multiMsToDs}
\title{Helper function to convert list of matrices to dissimilarity matrix}
\usage{
multiMsToDs(multiMsMat)
}
\arguments{
\item{Matrix}{list}
}
\value{
Dissimilarity Matrix
}
\description{
Helper function to convert list of matrices to dissimilarity matrix
}
\examples{
multiMsToDs(Matrix list)
}
