% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normVector.R
\name{normVector}
\alias{normVector}
\title{Helper function to normalize vector}
\usage{
normVector(Vec)
}
\arguments{
\item{Vector}{to be normalized}
}
\value{
Normalized vector
}
\description{
Helper function to normalize vector
}
\examples{
normVector(vector)
}
