% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packLoad.R
\name{packLoad}
\alias{packLoad}
\title{Helper function to load packages}
\usage{
packLoad(pkg)
}
\arguments{
\item{Package/list}{of packages}
}
\value{
None - packages are loaded
}
\description{
Helper function to load packages
}
\examples{
packLoad(list of packages)
}
