% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popBarPlotInd.R
\name{popBarPlotInd}
\alias{popBarPlotInd}
\title{Helper function to get the population bar plot for individuals}
\usage{
popBarPlotInd(toPlot_DF, highlightData = NULL, Population_Name,
  isName = FALSE)
}
\arguments{
\item{The}{dataframe, any highlighted data, the population name, and a name parameter (defaults to FALSE)}
}
\value{
Bar Plot
}
\description{
Helper function to get the population bar plot for individuals
}
\examples{
popBarPlotInd(Dataframe, highlighted data, population name)
}
