% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKinVis.R
\name{runKinVis}
\alias{runKinVis}
\title{Function that runs the application}
\usage{
runKinVis(dataPath)
}
\arguments{
\item{Absolute}{path to the data folder}
}
\value{
Nothing - executes the application
}
\description{
Function that runs the application
}
\examples{
runKinVis(C://Data)
}
