% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarDF.R
\name{stackedBarDF}
\alias{stackedBarDF}
\title{Helper function to generate a data frame for a stacked bar plot}
\usage{
stackedBarDF(Vs, maxKVal = 9)
}
\arguments{
\item{Vector, }{maximum k-value (defaults to 9)}
}
\value{
Dataframe for the plot
}
\description{
Helper function to generate a data frame for a stacked bar plot
}
\examples{
stackedBarDF(Vector, Maximum K Value)
}
