% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecEucDistances.R
\name{vecEucDistances}
\alias{vecEucDistances}
\title{Helper function to return distance matrix}
\usage{
vecEucDistances(vecVsList)
}
\arguments{
\item{List}{of multiple vectors}
}
\value{
Distance matrix, computed using Euclidean distance formula
}
\description{
Helper function to return distance matrix
}
\examples{
vecEucDistance(List of vectors)
}
