% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecEucDistCalc.R
\name{vecEucDistCalc}
\alias{vecEucDistCalc}
\title{Helper function to calculate Euclid distance between two vectors}
\usage{
vecEucDistCalc(vec1, vec2)
}
\arguments{
\item{Two}{vectors}
}
\value{
Euclidean distance between two input vectors
}
\description{
Helper function to calculate Euclid distance between two vectors
}
\examples{
vecEucDistCalc(vector1, vector2)
}
