% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecGetPopName.R
\name{vecGetPopName}
\alias{vecGetPopName}
\title{Helper function to get population names from the vector of file names}
\usage{
vecGetPopName(fileNameVec)
}
\arguments{
\item{Vector}{of file names}
}
\value{
Vector of population names
}
\description{
Helper function to get population names from the vector of file names
}
\examples{
vecGetPopName(vector of file names)
}
