% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecGetProcName.R
\name{vecGetProcName}
\alias{vecGetProcName}
\title{Helper function to get process name from the vector of file names}
\usage{
vecGetProcName(fileNameVec)
}
\arguments{
\item{Vector}{of file names}
}
\value{
Process names
}
\description{
Helper function to get process name from the vector of file names
}
\examples{
vecGetProcName(vector of file names)
}
