% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecToDataFrame.R
\name{vecToDataFrame}
\alias{vecToDataFrame}
\title{Helper function to convert a vector to a dataframe}
\usage{
vecToDataFrame(Vs, maxKSet)
}
\arguments{
\item{Vector, }{max K value}
}
\value{
Created data frame
}
\description{
Helper function to convert a vector to a dataframe
}
\examples{
vecToDataFrame(vector, max K)
}
