% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecToDFMDS.R
\name{vecToDFMDS}
\alias{vecToDFMDS}
\title{Helper function to convert a vector to a dataframe for MDS}
\usage{
vecToDFMDS(vecVsList)
}
\arguments{
\item{List}{of vectors}
}
\value{
Created data frame for the MDS Plot
}
\description{
Helper function to convert a vector to a dataframe for MDS
}
\examples{
vecToDFMDS(list of vectors)
}
