% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VecToFrq.R
\name{VecToFrq}
\alias{VecToFrq}
\title{Helper function to convert a vector to list of frequences}
\usage{
VecToFrq(vec, maxK = 9)
}
\arguments{
\item{Vector}{to convert}
}
\value{
Vector of frequencies
}
\description{
Helper function to convert a vector to list of frequences
}
\examples{
VecToFrq(vector, max K)
}
